chore: refactor example
